Move App main component into its own components/folder
