+++ src/components/App/App.jsx
@@ -1,10 +1,10 @@
 import { Routes, Route } from 'react-router-dom';
-import BookSpot from './routes/BookSpot';
-import Home from './routes/Home';
-import Categories from './routes/Categories';
-import Profile from './routes/Profile';
-import NotFound from './routes/NotFound';
-import './assets/styles/App.css';
+import BookSpot from '../../routes/BookSpot';
+import Home from '../../routes/Home';
+import Categories from '../../routes/Categories';
+import Profile from '../../routes/Profile';
+import NotFound from '../../routes/NotFound';
+import '../../assets/styles/App.css';
 
 const App = () => (
   <div className="App">
